simplify testing of device scan result parsing
